Fixes to room and Kevins room fully working

+++ Source/PyramidGame/Litetspel_Horus_Pyramid/Litetspel_Horus_Pyramid/Input_Mouse.cpp
@@ -66,7 +66,7 @@ bool Mouse::empty()
 void Mouse::onMove(MousePos pos)
 {
 	this->m_mousePos = pos;
-	this->m_events.push(MouseEvent(Event::MouseMove, pos.x*0.25, pos.y* 0.25));
+	this->m_events.push(MouseEvent(Event::MouseMove, (int)(pos.x*0.25), (int)(pos.y* 0.25)));
 }
 void Mouse::onRawMove(MousePos pos)
 {
